Add special-case handling for the description predicate.
Descriptions are really weird beasts compared to the rest of Prov-N, or my
understanding of the rest of prov-n at least.  I've added a rather hackish way
to produce RDF for descriptions but this needs revisited after the demo.

1. Make a new lexer state so we lex ident/num correctly in attributes.
2. Enumerate the predicate attributes separately, breaking apart the monolithic
PredicateAttr type.
3. Ensure the predicate 'object' is the named object and I'm not just using the
object field for the subject.  What does the prov spec say here?