Fix #5: Add more flexible exception handling.
@m271828 Notice I haven't had a chance to test this yet, but it seems
straight-forward.  The idea is that the old `titan` function had a built-in set
of exceptions it would catch and re-throw the others.  I've moved this rather
reasonable set of exceptions-to-handle into their own function,
`defaultErrorHandle`.  Users can either depend on this default or supply
an `ErrorHandle` value that indicates how they would prefer exceptions be treated.

In `titanWith` we always catch the `HttpException`s by converting them to an Either
type.  We then consult the provided error handle to decide what action to take.

Notice the default now includes a single retry on response-timeout exceptions,
so issue #5 should be handled (or silently ignored on double failure).
Our Trint code does not require any changes because it never used the return
value, which is the only externally visible change to the `titan` function.