Add a new fake classifier using aiogremlin
This is untested since the fake* infrastructure is rather dead, but the
aiogrmlin documents appear straight forward. Near as I can tell, you just use
the given AIO event to acquire futures that represent the result for an
asynchronous IO operation.  Since the demo doesn't really care about the async
part we just block till the result is available and print.