Batch insertions and vertex index via 'id' not 'label'.
The batch insertions are to avoid a very real issue with Gremlin-Server in which
it appears to block because "write buffer is full" then, because it is blocked,
time out on the current insertion.  Our alternative, untested, is to batch the
insertions and force some degree of synchronous operation.

Using 'id' instead of 'label' is all about performance ('id' is indexed so it
avoids the O(|V|) graph traversal).