Use 'id' instead of label.
By using an ElementIdStrategy with Titan we can use custom node identifiers.
With these, we don't need to match insertions with response id numbers and do
the whole memoization trick.  Instead, we can use the prov identifiers as node
identifiers.

This patch also tweaks the batch size in an attempt to play well with the
resources provided to our virtualized deployment used in development.