Remove dead edges before inseration.
We know we receive incomplete information and that leads to edges with
non-existent end points.  In titan if an edge is inserted with a non-existent
end it triggers a full database traversal, don't ask me why, so this really
really needs to be avoided.  In the single producer case we have all the
information locally and can simply do a lookup and a filter.  In a
multi-producer case something more complex will need to be dreamed up.