Rework Titan interface to use bindings.
"Gremlin server caches the script" - the "gremlin : 'g.addV(...)' part of each
JSON request, not the scripts/*.groovy files that I was thinking that sentence
meant.  So we should not use literals in the insertion command but instead use
variable names and bindings (and environment) since the bindings are not cached.