Add code paths for Trint to decode CDM bin.
I don't expect this to work as is.  For one, all these new code paths lack
testing (due to a lack of specification, understanding).  More-over, my reading
of Avro is that the schema is included in the encoding - the manner of its
inclusion is not in the spec I saw so working this out is going to be an
exploration.