Revert boolean change.
It seemed obvious we wanted `Nothing <- code` so the fall-through case was
explicit in what fell through to it, but the current GHC totality checker does
not know this function is total.  This strikes me as a bug - after de-sugaring
totality should be obvious as a case statement, but that's for another day.