Remove 'trust' and 'sensitivity'.
We previously tried to track down this information from the abstract object,
whicih held ProvenanceTagNodes.  Now such information is in a new node,
communicated separately, so we would have to buffer or lookup and modify past
nodes to ingest into the same format.  Instead we are doing what was publicly
stated all along - ignoring tag information completely (for now).