Change permissions and parameters to match /opt directory use.
I moved some services from $HOME/adapt to opt, since they are more of an
enterprise level things and this is in tune with how the ta3 vm in tc-in-a-box
is setup.  Feel free to revert if you disagree, anyone, with this rational. This
patch just corrects a few more locations where the directory structures
disagreed.