Add 'ta2' topic even on tc-in-a-box.
I see the old install/start_daemons.sh was removed as it was viewed as
duplicate.  It wasn't really duplicate, there was the adapt-in-a-box system and
the tc-in-a-box system which existed side by side in a poorly-named and
undocumented cludge for which, understandably, no one but me knew the layout.
Deleting the install/start_daemons.sh lead to "fixing" the Vagrantfile and now a
missing Kafka topic at run-time ("ta2" topic is used by Trint on Adapt-in-a-box
but is otherwise unused in tc-in-a-box).

To go with the flow while improving our layout this is the first of many patches
aimed at:

1. Moving The adapt-in-a-box Vagrantfile to the top-level.
2. Stop using file provisioning to distribute directories, which never works on
re-provision.
3. Use the auto-shared directory of (in guest) '/vagrant' to get components
4. Renaming or fully removing 'install' since the adapt-in-a-box installation
will happen from top-level.