Make Titan WS UUIDs unique.
The UUIDs are computed as hashes of the command being executed, which is unique
right?  Well, it was until we moved from full inlined statements to common
scripts + environments (using the 'bindings' field) - given that change we must
also include the environment in the UUID computation to again get unique UUIDs.