Only wait for 'REGISTERED' titan indexes when needed.
Titan indexes can be in one of a few states.  Initially they are 'installed' and
once the keys are available they transition to 'registered' and or 'enabled'
depending on specifics that seem poorly-documented.  Waiting for something to be
'registered' which is already 'enabled' costs 60 seconds then a timeout of the
wait routine.

N.B.  REINDEX already costs 3 minutes (no prior data, seems to be a timer as no
computation is occuring).  As a result, setting up a new system means that it is
unusable / will through connection exceptions for the first three minutes.