Fix #22: Make an explicit 'send finished' flag in Trint.
Trint -f: sends 0x01 (done) to ingestd (Kafka in-finsihed) which echos the
          message to PE.

Why not make the ingester receive a message on the data line (kafka 'in' topic),
as discussed? Because we can't make tc-in-a-box send us this message, which is
the end inter-op goal.  To work with Adapt-in-a-box as well as TC-in-a-box we
allow/require a human interaction where we push a button to kick off the rest of
the system.