Remove the call to GenerateSchema from titan script.
The script must _complete_ before the websocket interface is available, so
shelling out to python as part of the setup is not a functioning solution.  We
can either add the commands to the script as gremlin code or add them to
start_daemons.  Does gremlin have 'eval()'?  I'll look and use that if so.