Add EDGE_EVENT_AFFECTS_FILE
@dwa @m271828  For the record, I think including gremlin commands and regex
strings in the markdown is rather cludgy and hard to maintain.  If we want to
automate more of the process lets make tools that parse and work from the TA-3
provided Avro IDL.  As it stands the CDM on which our spec was based is
out-of-date and this causes breakages that are annoying to track down.  I
propose we consider retooling to lean more heavily on the TA-3 specification
when the dust settles.