Handle GID (list of int) properties correctly.
AHhh, so gremlin does not support lists as in:

    g.addV(label, 'foo', 'prop1', [1,2,3,4])

Which is infuriating.  But it does support lists as:

    g.addV(label, 'foo', 'prop1', 1, 'prop1', 2, 'prop1', 3, 'prop1', 4)

Which isn't a list as in a Java ArrayValue class, but it is a single Java
Integer.Class with a property cardinality > 1.  Not sure if order is retained,
which might be important, but for now this is what we're doing.