Insert 2 edges + 1 node for each CDM edge in one gremlin command.
See ticket GaloisInc/adapt#29 - we have an issue with asynchronously, via the
websocket, we are sending:

```
Add Node 5
Add Edge (Node 1) (Node 5)
Add Edge (Node 5) (Node 2)
```

This sequence seems to have a high failure rate for the edge insertions probably
because 'node 5' does not always exist prior to the edge insertion commands
being executed.  Instead, we can bundle this into a single command and at the
same time save on two lookups (finding the address of node 5 in each edge
insertion).  It remains to be seen if this is a fix, hence not closing the issue
just yet.