Remove reindexing entirely.
We don't need to reindex with zero data and, now that things are smoothing out,
we should never start up and use a database then add the index second.  I had
previously believed, based on Titan instructions, that reindexing is _the_ way
to transition an index to _ENABLED_.  This is not true.