Refactor adapt.groovy.
I've factored out the index insertion scripts into functions.  I believe these
functions are properly parameterized such that they will be useful for many of
the other scripts needed by the project.

IF YOU WANT TO ADD AN INDEX:

1. Decide if you want equality or richer search
2. If you want equality then use makeNodeIndex.
3. If you need search that is richer than equality (c.f. ElasticSearch) then use
makeElasticSearchIndex.
4. Provide the index name, which can be arbitrary but must not conflict with
other names and in the case of elastic search will be needed in the queries.
5. Provide a single index key.  You can make indexes over multiple keys but the
provided functions just encode the method for single-key indexes.
6. Provide the type, such as String.class, which will be indexed.