Ingestd exponential back-off & blocking finish.
Fix #81 and Fix #80

N.B. This patch reworks the database connection scope and exceptions.
One result is that each TA2 source has its own DB connection and its own limit
on the number of requests in flight.  This might be a bad way to go, but for the
first encounter it doesn't matter since there is only one source per ingestd
instance.

We made the retry back-off exponentially (4^age * 1 second, max of 1 minute but
die after age 3 right now).

If the finish signal is 0x02 (vs the norm of 0x01), do no propagate it from
ingestd until all incoming Kafka queues are empty.  The signal to SE is still 0x01.
`Trint -F` can be used to send the 0x02 signal.

We now delay longer in test framework because it has no feedback (unless you
count the later components, such as SE, blocking for the go signal).