First cut at boostrap/install partition.
The BBN machines are not vagrant deployed so we can't use bootstrap as-is to
install/reinstall the system.  The 'bootstrap.sh' script aim to be the same old
same old while the 'install.sh' script should be useful for (re)installing on
the BBN or other non-vagrant or non-git-cloning systems.