Fix permissions after generating schema.
I don't see any good reason sudo is needed on the GenerateSchema - I suspect
this was an issue with the 'darpa' user (vs 'vagrant') and not necessary.
Either way, this fix suffices but we need to clean this script at some point
and, sadly, agree on it's whole goal since those have expanded.