Order inputs to reduce variety of scripts.
Since gremlin server caches script evaluations it is greatly in our interest to
reduce the number of scripts even in the bulk case.  This can be accomplished by
either reducing the bulk load (not a good option) or by ensuring even when
batching the insertions we somehow use one of a small set of scripts.  To that
end, we sort of inputs into categories that will result in scripts of identical
forms.