adapt-tester auto-updates
The adapt-tester now compares a hash of itself to a hash on adapt.galois.com to determine if
it is outdated. If that is the case, it downloads the newer version of itself (and checks the
hash in case the download was corrupted). In either case, once it is on the newest version of
itself, it downloads the newest adapt.jar (checking a hash again) and runs the acceptance app.