Fix for gremlin tinkergraph
While the version in the previous commit would work when run from sbt (in dev and accept mode), it
would not work when assembled into a fat-jar (in either dev or accept mode). The error is (to the
best of my knowledge) related to some runtime-reflection shenanigans Groovy was doing in
GremlinGroovyScriptEngine.

After having tried several things, I've currently settled on querying the graph directly (not via
the groovy engine). We should consider not using GremlinGroovyScriptEngine altogether and writing
our own parser for queries.