Add an AST for traversals
This commit introduces a series of types to represent an AST for traversals. Along with this type
is functionality for parsing a traversal from a string, and running it on a graph. This allows us to
manipulate queries as first class objects (and gives us a way of bypassing groovy reflection which
is causing problems in fat jars).