Hack to get CDM14 working
Here is the problem: it appears that avro checks the packages when deserializing, so it is important
that we not change _anything_ about the avro specification. However, in an effort to remain sane, we
can

  * still move the generated `com.bbn.tc.schema.avro._` files into the `com.bbn.tc.schema.avro.cdm14._`
    _directory_ (but still have them polluting the more general namespace).
  * set up a package object for `com.bbn.tc.schema.avro.cdm14` that has type aliases so that we can
    still refer to `cdm14.AbstractObject` (for example).

It is my impression this is the (currently) least intrusive solution. I can now ingest/run tests on
our only CDM14 sample data file.

CAVEAT: I'm on a very poor internet connection, so I wasn't able to download the avro tools - I just
manually 'sed'ed the package declarations in the generated cdm14 java files (and changed the avro
spec). We should regenerate the Java files automatically and make sure they now end up in the right
package.