Switch everything to CDM18, reads CDM17 files too
Internally, everything (ER, anomaly, writing to DB) works over CDM18.
I have not accounted for changes in semantics - some of what is being
done could be making no sense.

When reading from a file, and if reading as CDM18 fails, we attempt
to read as CDM17 and map that into CDM18 (see the `Cdm17to18` object
in the `cdm18` package for what that conversion looks like).

This lets us continue to ingest old CDM17 files.