Add Fridge, expire events
Before this commit, event merging was buggy: we could find ourselves holding on to chains of events
that would grow forever, and we would not emit a resolved event until we _knew_ the next event
wouldn't be merged into it (so that means the last events for every process were never emitted).

Now: we keep track of time in the stream and, expire events if either too much time has passed since
the last event, or the chain of merged events is too long. The logic behind what to expire next is
encapsulated in the 'Fridge' data type.