"has('uuid',<lit-uuid>)" queries check all namespaces
This is a but more involved than appears at first sight: we have to
serialize to disk the namespaces we've encountered (so that we know
how to rewrite the queries even if just running the UI).

This is done with a `namespaces` file in the neo4j DB folder (which
is written to on shutdown and read on startup).