Option for keeping event uuids out of UuidRemapper
This is based on the assumption that edges never point to events. Given that this
assumption can hardly be expected to hold up in general, it is hidden behind the
'adapt.adm.ignoreeventremaps' flag.

On Cadets E2 bovia, this resulted in:

  * ~1.6x speedup
  * ~14x reduction in the size of 'cdm2adm'
  * ~8x reduction in the size of 'seenNodes'

Obviously, mileage will vary with the shape of the dataset.